Updated the flow

+++ src/rs-ec.cc
@@ -9,16 +9,19 @@
 
 USING_YOSYS_NAMESPACE
 PRIVATE_NAMESPACE_BEGIN
-
+namespace fs = std::filesystem;
 #define MODE_BITS_OUTPUT_SELECT_WIDTH 3
 #define GENESIS_3_DIR genesis3
 #define XSTR(val) #val
 #define STR(val) XSTR(val)
 #define GET_FILE_PATH(tech_dir,file) " +/rapidsilicon/" STR(tech_dir) "/" STR(file)
+#define GET_FILE_PATH_SEC_MODEL(tech_dir,file) "/rapidsilicon/" STR(tech_dir) "/" STR(file)
 #define SEC_SIM_CELLS SEC_MODELS/simcells.v
 #define SEC_SIM_LIB SEC_MODELS/simlib.v
+#define SEC_DFFRE_blif SEC_MODELS/DFFRE.blif
 
 bool is_genesis3;
+bool isSequential =false;
 string current_stage;
 string previous_state = "";
 int sec_counter = 0;
@@ -49,6 +52,26 @@ struct RsSECWorker
         }
     };
     
+    struct Entry_ff_sync {
+        RTLIL::IdString type;
+        int inCLK;
+        int inEN;
+        int inSET;
+        int inCLR;
+        int inD;
+        int outQ;
+        int clk_polarity;
+        int srst_polarity;
+        int srst_value;
+        // Overload == operator to compare two Entry objects
+        bool operator==(const Entry_ff_sync& other) const {
+            return type == other.type && inCLK == other.inCLK && inEN == other.inEN && inSET == other.inSET && \
+            inCLR == other.inCLR && inD == other.inD && outQ == other.outQ && clk_polarity == other.clk_polarity &&\
+            srst_polarity == other.srst_polarity && srst_value == other.srst_value;
+        }
+        
+    };
+    
     struct CompareEntry {
         bool operator()(const Entry& a, const Entry& b) const {
             if (a.type != b.type) return a.type < b.type;
@@ -60,6 +83,21 @@ struct RsSECWorker
         }
     };
     
+    struct CompareEntry_ff_sync {
+        bool operator()(const Entry_ff_sync& a, const Entry_ff_sync& b) const {
+            if (a.type != b.type) return a.type < b.type;
+            if (a.inEN != b.inEN) return a.inEN < b.inEN;
+            if (a.inSET != b.inSET) return a.inSET < b.inSET;
+            if (a.inCLR != b.inCLR) return a.inCLR < b.inCLR;
+            if (a.inD != b.inD) return a.inD < b.inD;
+            if (a.outQ != b.outQ) return a.outQ < b.outQ;
+            if (a.clk_polarity != b.clk_polarity) return a.clk_polarity < b.clk_polarity ;
+            if (a.srst_polarity != b.srst_polarity) return a.srst_polarity < b.srst_polarity;
+            if (a.srst_value != b.srst_value) return a.srst_value < b.srst_value;
+            return a.inCLK < b.inCLK;
+        }
+    };
+    
     void addEntry(std::vector<Entry>& entries, std::set<Entry, CompareEntry>& uniqueSet, const Entry& entry) {
         if (uniqueSet.find(entry) == uniqueSet.end()) {
             uniqueSet.insert(entry);
@@ -67,13 +105,139 @@ struct RsSECWorker
         }
     }
     
+    void addEntry_ff_sync(std::vector<Entry_ff_sync>& entries, std::set<Entry_ff_sync, CompareEntry_ff_sync>& uniqueSet, const Entry_ff_sync& entry) {
+        if (uniqueSet.find(entry) == uniqueSet.end()) {
+            uniqueSet.insert(entry);
+            entries.push_back(entry);
+        }
+    }
+    
+    void add_ff_sync (RTLIL::Design *design, const std::vector<Entry_ff_sync>& entries){
+        for (auto entry : entries){
+            if (entry.type.in(ID($ff), ID($dff), ID($dffe), ID($dffsr),ID($dffsre),ID($_DFF_P_),ID($_DFF_N_),ID($dff))){
+                log("%s: %d, %d, %d, %d, %d, %d\n",entry.type.c_str(),entry.inCLK, entry.inEN, entry.inSET, entry.inCLR, entry.inD, entry.outQ);
+                std::string cell_type_str = entry.type.str();
+                string  mod_name = cell_type_str + "_" + std::to_string(entry.inCLK) + "_" + std::to_string(entry.inEN) + "_" + std::to_string(entry.inD)+"_"+std::to_string(entry.outQ);
+                log("Module = %s\n",mod_name.c_str());
+                design->addModule(RTLIL::escape_id(mod_name));
+
+                for (auto &module : design->selected_modules()) {
+                    if (module->name == RTLIL::escape_id(mod_name) && entry.type== ID($dff)){
+                        SigSpec CLK;
+                        SigSpec D; 
+                        SigSpec Q;  
+
+                        CLK = module->addWire(RTLIL::escape_id("CLK"), entry.inCLK);
+                        D   = module->addWire(RTLIL::escape_id("D"), entry.inD);
+                        Q   = module->addWire(RTLIL::escape_id("Q"), entry.outQ);
+                        for (auto wire : module->wires()){
+                            if (wire->name == "\\CLK") {
+                                wire->port_input = true;
+                                wire->port_id=1;
+                            }
+                            if (wire->name == "\\D") {
+                                wire->port_input = true;
+                                wire->port_id=2;
+                            }
+                            if (wire->name == "\\Q") {
+                                wire->port_output = true;
+                                    wire->port_id=3;
+                            }
+                        }
+                        module->ports.push_back(RTLIL::escape_id("CLK"));
+                        module->ports.push_back(RTLIL::escape_id("D"));
+                        module->ports.push_back(RTLIL::escape_id("Q"));
+                        if (entry.type == "$dff")
+                            module->addDff(RTLIL::escape_id(mod_name), CLK, D, Q,true);
+                    }
+                    else if (module->name == RTLIL::escape_id(mod_name) && entry.type== ID($sdff)){
+                        SigSpec CLK;
+                        SigSpec SRST;
+                        SigSpec D; 
+                        SigSpec Q;  
+                        bool clk_polarity= entry.clk_polarity;
+                        bool srst_polarity = entry.srst_polarity;
+                        CLK = module->addWire(RTLIL::escape_id("CLK"), entry.inCLK);
+                        SRST = module->addWire(RTLIL::escape_id("SRST"), entry.inCLR);
+                        D   = module->addWire(RTLIL::escape_id("D"), entry.inD);
+                        Q   = module->addWire(RTLIL::escape_id("Q"), entry.outQ);
+                        for (auto wire : module->wires()){
+                            if (wire->name == "\\CLK") {
+                                wire->port_input = true;
+                                wire->port_id=1;
+                            }
+                            if (wire->name == "\\SRST") {
+                                wire->port_input = true;
+                                wire->port_id=2;
+                            }
+                            if (wire->name == "\\D") {
+                                wire->port_input = true;
+                                wire->port_id=3;
+                            }
+
+                            if (wire->name == "\\Q") {
+                                wire->port_output = true;
+                                    wire->port_id=4;
+                            }
+                        }
+                        module->ports.push_back(RTLIL::escape_id("CLK"));
+                        module->ports.push_back(RTLIL::escape_id("SRST"));
+                        module->ports.push_back(RTLIL::escape_id("D"));
+                        module->ports.push_back(RTLIL::escape_id("Q"));
+                        module->addSdff(RTLIL::escape_id(mod_name), CLK, SRST, D, Q,entry.srst_value,clk_polarity,srst_polarity);
+                    }
+                    else if (module->name == RTLIL::escape_id(mod_name) && (entry.type== ID($_DFF_P_) || entry.type== ID($_DFF_N_))){
+                        SigSpec C;
+                        SigSpec D; 
+                        SigSpec Q;  
+
+                        C = module->addWire(RTLIL::escape_id("C"), entry.inCLK);
+                        D   = module->addWire(RTLIL::escape_id("D"), entry.inD);
+                        Q   = module->addWire(RTLIL::escape_id("Q"), entry.outQ);
+                        for (auto wire : module->wires()){
+                            if (wire->name == "\\C") {
+                                wire->port_input = true;
+                                wire->port_id=1;
+                            }
+                            if (wire->name == "\\D") {
+                                wire->port_input = true;
+                                wire->port_id=2;
+                            }
+                            if (wire->name == "\\Q") {
+                                wire->port_output = true;
+                                    wire->port_id=3;
+                            }
+                        }
+                        module->ports.push_back(RTLIL::escape_id("C"));
+                        module->ports.push_back(RTLIL::escape_id("D"));
+                        module->ports.push_back(RTLIL::escape_id("Q"));
+                        if (entry.type == "$_DFF_P_")
+                            module->addDffGate(RTLIL::escape_id(mod_name), C, D, Q,true);
+                        else
+                            module->addDffGate(RTLIL::escape_id(mod_name), C, D, Q,false);
+                    }
+                }
+                Pass::call(design, stringf("hierarchy -top %s",mod_name.c_str()));
+
+                // Pass::call(design, "stat");
+                Pass::call(design, "proc");
+                Pass::call(design, "flatten");
+                Pass::call(design, "opt_expr");
+                Pass::call(design, "techmap");
+                Pass::call(design, "opt_clean");
+                Pass::call(design, stringf("write_blif %s.blif",mod_name.c_str()));
+            }
+        }
+    }
+    
     void add_A_B_Y_S (RTLIL::Design *design, const std::vector<Entry>& entries){
         for (auto entry : entries){
             log("%s: %d, %d, %d, %d\n",entry.type.c_str(),entry.inA, entry.inB, entry.outY, entry.signed_opr);
             string  mod_name =  entry.type+ "_" + std::to_string(entry.inA) + "_" + std::to_string(entry.inB) + "_" + std::to_string(entry.outY)+"_"+std::to_string(entry.inS);
+            
             log("Module = %s\n",mod_name.c_str());
             design->addModule(RTLIL::escape_id(mod_name));
-            
+
             for (auto &module : design->selected_modules()) {
                 if (module->name == RTLIL::escape_id(mod_name)){
                     SigSpec A;
@@ -151,9 +315,10 @@ struct RsSECWorker
         if (!gen_net)
             Pass::call(design, "design -save original");
         Pass::call(design,"opt_clean");
-        
+
         for(auto& modules : design->selected_modules()){
             for (auto &cell : modules->selected_cells()) {
+
                 if (cell->type.in(ID($add),ID($xor),ID($and),ID($or))){
                     string cell_type = log_id(cell->type);
                     RTLIL::IdString cell_name = cell_type + "_" + \
@@ -181,6 +346,30 @@ struct RsSECWorker
                     blif_models.insert(cell_name);
                     cell->type = cell_name;
                 }
+                if (cell->type.in(ID($dff),ID($sdff))){
+                    string cell_type = log_id(cell->type);
+                    RTLIL::IdString cell_name = cell_type + "_" + \
+                    std::to_string(GetSize(cell->getPort(ID::CLK))) + "_" +"0" + "_" + \
+                    std::to_string(GetSize(cell->getPort(ID::D))) + "_" + \
+                    std::to_string(GetSize(cell->getPort(ID::Q)));
+                    blif_models.insert(cell_name);
+                    cell->type = cell_name;
+                }
+                if (cell->type.in(ID($_DFF_P_),ID($_DFF_N_))){
+                    string cell_type = log_id(cell->type);
+                    RTLIL::IdString cell_name = cell_type + "_" + \
+                    std::to_string(GetSize(cell->getPort(ID::C))) + "_" +"0" + "_" + \
+                    std::to_string(GetSize(cell->getPort(ID::D))) + "_" + \
+                    std::to_string(GetSize(cell->getPort(ID::Q)));
+                    blif_models.insert(cell_name);
+                    cell->type = cell_name;
+                }
+                if (cell->type == RTLIL::escape_id("\\DFFRE")){
+                    RTLIL::IdString cell_name = cell->type; 
+                    blif_models.insert(cell->type);
+                    
+                }
+
             }
         }
     }
@@ -195,25 +384,49 @@ struct RsSECWorker
         string model_file_name,netlist_name;
 
         for (auto blif_model : blif_models){
-            model_file_name = currentPathStr + "/" + log_id(blif_model) +".blif";
-            
-            std::ifstream inputFile(model_file_name);
-            if (!inputFile.is_open()) 
-                log_error("Failed to open the input file %s\n",model_file_name.c_str());
 
-            netlist_name = currentPathStr + "/" + current_netlist_name + ".blif";
-            std::ofstream outputFile(netlist_name, std::ios::app);
-            if (!outputFile.is_open()) 
-                log_error("Failed to open the output file %s\n",netlist_name.c_str());
+            if(blif_model == RTLIL::escape_id("DFFRE")){
+                std::string model_file_name = GET_FILE_PATH_SEC_MODEL(GENESIS_3_DIR, SEC_DFFRE_blif);
+                string plugin_dir = proc_self_dirname() + "../share/yosys/";
+                model_file_name = plugin_dir + model_file_name;
+                std::ifstream inputFile(model_file_name);
+                if (!inputFile.is_open()) 
+                    log_error("Failed to open the input file %s\n",model_file_name.c_str());
+
+                netlist_name = currentPathStr + "/" + current_netlist_name + ".blif";
+                std::ofstream outputFile(netlist_name, std::ios::app);
+                if (!outputFile.is_open()) 
+                    log_error("Failed to open the output file %s\n",netlist_name.c_str());
             
-            std::string line;
-            while (std::getline(inputFile, line)) {
-                outputFile << line << std::endl;
+                std::string line;
+                while (std::getline(inputFile, line)) {
+                    outputFile << line << std::endl;
+                }
+
+                inputFile.close();
+                outputFile.close();
+
+            }else {
+                model_file_name = currentPathStr + "/" + log_id(blif_model) +".blif";
+            
+                std::ifstream inputFile(model_file_name);
+                if (!inputFile.is_open()) 
+                    log_error("Failed to open the input file %s\n",model_file_name.c_str());
+
+                netlist_name = currentPathStr + "/" + current_netlist_name + ".blif";
+                std::ofstream outputFile(netlist_name, std::ios::app);
+                if (!outputFile.is_open()) 
+                    log_error("Failed to open the output file %s\n",netlist_name.c_str());
+            
+                std::string line;
+                while (std::getline(inputFile, line)) {
+                    outputFile << line << std::endl;
+                }
+
+                std::filesystem::remove(model_file_name);
+                inputFile.close();
+                outputFile.close();
             }
-
-            std::filesystem::remove(model_file_name);
-            inputFile.close();
-            outputFile.close();
                 
         }
     }
@@ -285,8 +498,8 @@ struct RsSECWorker
        return false;
     }
 
-    void run_sec(RTLIL::Design *design){
-        bool isSequential = design_has_clk(design);
+    void run_sec(bool Sequential){
+        
         char current_netlist_name[528];
         char previous_netlist_name[528];
         char resName[528];
@@ -295,7 +508,7 @@ struct RsSECWorker
         sprintf(resName, "%03d_%s.res", (sec_counter), current_stage.c_str());
         string cmd;
         string abc_script = "abc.scr";
-        if (isSequential){
+        if (Sequential){
             cmd = std::string("dsec ") + current_netlist_name + std::string(" ") +  previous_netlist_name +std::string("; quit;");
         }
         else{
@@ -313,7 +526,7 @@ struct RsSECWorker
         std::string command = abc_exe_dir + std::string(" -f abc.scr > ") + resName;
         command += std::string(" ; tail -n 2 ") + resName;
         
-        if (isSequential) {
+        if (Sequential) {
             log("DSEC : \n");
         } else {
             log("CEC : \n");
@@ -330,6 +543,7 @@ struct RsSECWorker
     void gen_netlist(RTLIL::Design *design){
         std::string simcells = GET_FILE_PATH(GENESIS_3_DIR, SEC_SIM_CELLS);
         std::string simlib = GET_FILE_PATH(GENESIS_3_DIR, SEC_SIM_LIB);
+        
         Module* topModule = design->top_module();
 
         const char* topName = topModule->name.c_str();
@@ -354,27 +568,43 @@ struct RsSECWorker
     }
 
     void run_scr (RTLIL::Design *design) {
-        
+        isSequential = design_has_clk(design);
         Pass::call(design, "design -save original");
         if (gen_net)
             gen_netlist(design);
 
         std::vector<Entry> entries;
         std::set<Entry, CompareEntry> uniqueSet;
+
+        std::vector<Entry_ff_sync> entries_ff;
+        std::set<Entry_ff_sync, CompareEntry_ff_sync> uniqueSet_ff;
         
         for (auto cell : design->top_module()->cells()) {
             if (cell->type.in(ID($add), ID($sub), ID($div), ID($mod), ID($divfloor), ID($modfloor), ID($pow), ID($mul),\
                 ID($xor),ID($or),ID($or),ID($and),ID($nor),ID($nand))){
                 addEntry(entries,uniqueSet,{log_id(cell->type),GetSize(cell->getPort(ID::A)),GetSize(cell->getPort(ID::B)),GetSize(cell->getPort(ID::Y)),false,0});
             }
-            if (cell->type.in(ID($reduce_xor),ID($not))){
+            if (cell->type.in(ID($reduce_xor),ID($not), ID($logic_not))){
                 addEntry(entries,uniqueSet,{log_id(cell->type), GetSize(cell->getPort(ID::A)), 0, GetSize(cell->getPort(ID::Y)), false,0});
             }
             if (cell->type.in(ID($mux))){
                 addEntry(entries,uniqueSet,{log_id(cell->type), GetSize(cell->getPort(ID::A)), 0, GetSize(cell->getPort(ID::Y)), false, GetSize(cell->getPort(ID::S))});
             }
+            if (cell->type.in(ID($shl), ID($shr), ID($sshl), ID($sshr))){
+                addEntry(entries,uniqueSet,{log_id(cell->type),GetSize(cell->getPort(ID::A)),GetSize(cell->getPort(ID::B)),GetSize(cell->getPort(ID::Y)),false,0});
+            }
+            if (cell->type.in(ID($dff))){
+                addEntry_ff_sync(entries_ff,uniqueSet_ff,{log_id(cell->type), GetSize(cell->getPort(ID::CLK)), 0, 0, 0,GetSize(cell->getPort(ID::D)), GetSize(cell->getPort(ID::Q)), false,false,0});
+            }
+            if (cell->type.in(ID($_DFF_P_),ID($_DFF_N_))){
+                addEntry_ff_sync(entries_ff,uniqueSet_ff,{log_id(cell->type), GetSize(cell->getPort(ID::C)), 0, 0, 0,GetSize(cell->getPort(ID::D)), GetSize(cell->getPort(ID::Q)),false, false, 0});
+            }
+            if (cell->type.in(ID($sdff))){
+                addEntry_ff_sync(entries_ff,uniqueSet_ff,{log_id(cell->type), GetSize(cell->getPort(ID::CLK)), 0, 0, GetSize(cell->getPort(ID::SRST)),GetSize(cell->getPort(ID::D)), GetSize(cell->getPort(ID::Q)),cell->getParam(ID::CLK_POLARITY).as_int(),cell->getParam(ID::SRST_POLARITY).as_int(),cell->getParam(ID::SRST_VALUE).as_int()});
+            }
         }
-        
+        add_ff_sync(design,entries_ff);
+       
         add_A_B_Y_S(design,entries);
     #if 1
         // Load previous step to show the 'stat'
@@ -413,7 +643,7 @@ struct RsSECWorker
         create_blif_models(design, blif_models);
         
         if (current_stage != "" && previous_state != "")
-            run_sec(design);
+            run_sec(isSequential);
         
         Pass::call(design, "design -load original");
         Pass::call(design,"design -save previous");
